Jung dae seung

+++ B2B-Semi/src/main/resources/static/js/book/bookList.js
@@ -270,7 +270,6 @@ let currentRating = 0;
 
 // 별점 선택시
 function updateStars(rating) {
-	console.log(rating);
 	const starRating = rating / 2; // 10점 만점을 5점 만점으로 변환
 	stars.forEach(star => {
 		// 기존의 클래스 삭제
@@ -532,6 +531,12 @@ function setupEditRatingLogic(reviewItem, currentStarPoint) {
         editCurrentRating = isLeftHalf ? (rating * 2 - 1) : rating * 2;
         editSelectedRating.textContent = editCurrentRating;
     });
+	
+	// 마우스가 떠났을때 
+	editRatingSelect.addEventListener('mouseleave', () => { 
+	    updateEditStars(editCurrentRating); 
+	    editSelectedRating.textContent = editCurrentRating; 
+	});
 }
 
 // 리뷰 저장 함수
@@ -585,7 +590,6 @@ function saveReview(reviewItem) {
 // 리뷰 삭제
 function deleteReview(reviewItem) {
 	
-	console.log(reviewItem);
 	const editedRating = reviewItem.querySelector('.review-actions').textContent;
 	const bookId = document.querySelector('#selectBookId').value;
 	
@@ -595,6 +599,11 @@ function deleteReview(reviewItem) {
 		"starPoint": parseFloat(editedRating),
         "memberNo": loginMember.memberNo
     };
+	
+	if (!confirm("해당 리뷰를 삭제 하시겠습니까?")) {
+		alert("취소됬습니다.");
+		return;
+	}
 	
     fetch("/book/deleteReview", {
         method: "DELETE",
